Conflicts:
	examples/config_motor/motor_configs/motor_config_Maxon_251601.h

+++ module_controllers_lib/include/controllers_lib.h
@@ -12,12 +12,12 @@
  * @brief Structure type to set the parameters of the PID controller.
  */
 typedef struct {
-    float Kp;
-    float Ki;
-    float Kd;
-    float integral_limit;
-    float integral;
-    float actual_value_1n;
+    double Kp;
+    double Ki;
+    double Kd;
+    double integral_limit;
+    double integral;
+    double actual_value_1n;
     int T_s;    //Sampling-Time in microseconds
 } PIDparam;
 
@@ -73,12 +73,12 @@ typedef struct {
  * @brief Structure type to set the parameters of the integral optimum position controller.
  */
 typedef struct {
-    float Kp;
-    float Ki;
-    float Kd;
-    float integral_limit;
-    float integral;
-    float actual_value_1n;
+    double Kp;
+    double Ki;
+    double Kd;
+    double integral_limit;
+    double integral;
+    double actual_value_1n;
     int T_s;    //Sampling-Time in microseconds
 } integralOptimumPosControllerParam;
 
@@ -111,7 +111,7 @@ void pid_init(PIDparam &param);
  * @param input, sample-time in us (microseconds).
  * @param the parameters of the PID controller
  */
-void pid_set_parameters(float Kp, float Ki, float Kd, float integral_limit, int T_s, PIDparam &param);
+void pid_set_parameters(double Kp, double Ki, double Kd, double integral_limit, int T_s, PIDparam &param);
 
 
 /**
@@ -122,7 +122,7 @@ void pid_set_parameters(float Kp, float Ki, float Kd, float integral_limit, int 
  * @param input, sample-time in us (microseconds).
  * @param the parameters of the PID controller
  */
-float pid_update(float desired_value, float actual_value, int T_s, PIDparam &param);
+double pid_update(double desired_value, double actual_value, int T_s, PIDparam &param);
 
 
 /**
